requirements.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
faiss-cpu==1.13.2
fastapi==0.109.2
filelock==3.20.3
fonttools==4.61.1
frozendict==2.4.7
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
llvmlite==0.46.0b1
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multitasking==0.0.12
networkx==3.6.1
numba==0.63.0b1
numpy==2.4.1
packaging==25.0
pandas==2.3.3
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
protobuf==6.33.4
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.9.0
shap==0.50.0
six==1.17.0
slicer==0.0.8
soupsieve==2.8.1
starlette==0.36.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.27.1
websockets==16.0
xgboost==3.1.3
yfinance==1.0
rank-bm25==0.2.2
